+++ madigan_cpp/Synth.cpp
@@ -1,28 +1,28 @@
 #include "Synth.h"
 
 
-std::vector<float> Synth::generate(){
+std::vector<double> Synth::generate(){
   return generator->next();
 }
 
-State Synth::preprocess(std::vector<float> prices){
-  std::vector<float> normedPortfolio = portfolioNorm();
+State Synth::preprocess(std::vector<double> prices){
+  std::vector<double> normedPortfolio = portfolioNorm();
 
   return State(prices, normedPortfolio);
 }
 
-float Synth::equity(){
-  float _sum(0.);
+double Synth::equity(){
+  double _sum(0.);
   for (int i=0; i != nAssets; i++){
     _sum += (_currentPrices[i] * _portfolio[i]);
   }
   return _cash + _sum;
 }
 
-float Synth::availableMargin(){
-  float _sum(0.);
+double Synth::availableMargin(){
+  double _sum(0.);
   for (int i=0; i != nAssets; i++){
-    float port = _portfolio[i];
+    double port = _portfolio[i];
     if(port < 0.){
       _sum += (_currentPrices[i] * port);
     }
@@ -30,7 +30,7 @@ float Synth::availableMargin(){
   return _cash + _sum;
 }
 
-bool Synth::checkRisk(int assetId, float amount){
+bool Synth::checkRisk(int assetId, double amount){
   if (amount == 0.) return false;
   if (amount > 0.){
     if (_portfolio[assetId] < 0.) return true;
@@ -42,8 +42,8 @@ bool Synth::checkRisk(int assetId, float amount){
   }
 }
 
-void Synth::transaction(int assetId, float amount, float transPrice, float transCost){
-  float unitsToExchange = amount / transPrice;
+void Synth::transaction(int assetId, double amount, double transPrice, double transCost){
+  double unitsToExchange = amount / transPrice;
   _cash -= (amount + transCost);
   _portfolio[assetId] += unitsToExchange;
 }
@@ -55,28 +55,28 @@ envOutput Synth::step(std::vector<int> actions){
 
   if (equity() <= 0.){
     bool done = true;
-    float reward = 0.;
+    double reward = 0.;
     Info info("BLOWNOUT",
-              std::vector<float>(nAssets, 0.),
-              std::vector<float>(nAssets, 0.));
-    std::vector<float> data = generate();
+              std::vector<double>(nAssets, 0.),
+              std::vector<double>(nAssets, 0.));
+    std::vector<double> data = generate();
     _currentPrices = data;
     State state(_currentPrices, portfolioNorm());
     return std::make_tuple(state, reward, done, info);
   }
   if (availableMargin() <= _maintenanceMargin){
     bool done = true;
-    float reward = 0.;
+    double reward = 0.;
     Info info("MARGINCALL",
-              std::vector<float>(nAssets, 0.),
-              std::vector<float>(nAssets, 0.));
-    std::vector<float> data = generate();
+              std::vector<double>(nAssets, 0.),
+              std::vector<double>(nAssets, 0.));
+    std::vector<double> data = generate();
     _currentPrices = data;
     State state(_currentPrices, portfolioNorm());
     return std::make_tuple(state, reward, done, info);
   }
 
-  float currentEquity = equity();
+  double currentEquity = equity();
   Info info;
 
   if (discreteActions){
@@ -88,8 +88,8 @@ envOutput Synth::step(std::vector<int> actions){
 
   _currentPrices = generate();
 
-  float nextEquity = equity();
-  float immediateReturn = (nextEquity/currentEquity) - 1.;
+  double nextEquity = equity();
+  double immediateReturn = (nextEquity/currentEquity) - 1.;
 
   State state = preprocess(_currentPrices);
 
@@ -100,14 +100,14 @@ envOutput Synth::step(std::vector<int> actions){
 
 
 Info Synth::stepDiscrete(std::vector<int> actions){
-  std::vector<float> transPrices;
-  std::vector<float> transCosts;
+  std::vector<double> transPrices;
+  std::vector<double> transCosts;
   for(int i=0; i != nAssets; i++){
-    float price = _currentPrices[i];
+    double price = _currentPrices[i];
     int lot =  lotUnitValue * (actions[i] - discreteActionAtomsShift);
-    float slippage = price * slippagePct * (lot<0? -1: 1);
-    float transPrice = price + slippage;
-    float amount = lot * transPrice;
+    double slippage = price * slippagePct * (lot<0? -1: 1);
+    double transPrice = price + slippage;
+    double amount = lot * transPrice;
     if (checkRisk(i, amount)){
       transaction(i, amount, transPrice, transactionCost);
       transPrices.push_back(transPrice);
